added shadow sprites and updated the door sprites to fit the artstyle
added three shadow sizes in the sprite tab of the art folder, and i also changed the door sprites to be able to move and make sense for the artstyle. i also changed some metadata for some sprites, i will continue to do that as our pixel size isnt consistent right now.